refactor(s3): tidy up props which aren't needed
* pull out s3file and add some interface checks
