+++ s3iofs.go
@@ -5,7 +5,6 @@ import (
 	"io/fs"
 	"os"
 	"path"
-	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
@@ -48,7 +47,6 @@ func (s3fs *S3FS) Open(name string) (fs.File, error) {
 			name:   name,
 			bucket: s3fs.bucket,
 			mode:   fs.ModeDir,
-			isDir:  true,
 		}, nil
 	}
 
@@ -131,7 +129,6 @@ func (s3fs *S3FS) ReadDir(name string) ([]fs.DirEntry, error) {
 			name:   dir,
 			bucket: s3fs.bucket,
 			mode:   fs.ModeDir,
-			isDir:  true,
 		})
 	}
 
@@ -160,7 +157,6 @@ func (s3fs *S3FS) stat(name string) (fs.FileInfo, error) {
 			name:   name,
 			bucket: s3fs.bucket,
 			mode:   fs.ModeDir,
-			isDir:  true,
 		}, nil
 	}
 
@@ -181,7 +177,6 @@ func (s3fs *S3FS) stat(name string) (fs.FileInfo, error) {
 			name:   name,
 			bucket: s3fs.bucket,
 			mode:   fs.ModeDir,
-			isDir:  true,
 		}, nil
 	}
 
@@ -209,71 +204,4 @@ func (s3fs *S3FS) openDirectory(name string) (fs.File, error) {
 	}
 
 	return nil, &fs.PathError{Op: "open", Path: name, Err: fs.ErrNotExist}
-}
-
-type s3File struct {
-	name    string
-	bucket  string
-	res     *s3.GetObjectOutput
-	size    int64
-	mode    fs.FileMode
-	modTime time.Time // zero value for directories
-	isDir   bool
-}
-
-func (s3f *s3File) Stat() (fs.FileInfo, error) {
-	return s3f, nil
-}
-
-func (s3f *s3File) Info() (fs.FileInfo, error) {
-	return s3f, nil
-}
-
-func (s3f *s3File) Read(p []byte) (int, error) {
-	if s3f.isDir {
-		return 0, &fs.PathError{Op: "read", Path: s3f.name, Err: errors.New("is a directory")}
-	}
-	return s3f.res.Body.Read(p)
-}
-
-func (s3f *s3File) Close() error {
-	if s3f.isDir {
-		return nil // NOOP for directories
-	}
-	return s3f.res.Body.Close()
-}
-
-// base name of the file
-func (s3f *s3File) Name() string {
-	return s3f.name
-}
-
-// length in bytes for regular files; system-dependent for others
-func (s3f *s3File) Size() int64 {
-	return s3f.size
-}
-
-// file mode bits
-func (s3f *s3File) Mode() fs.FileMode {
-	return s3f.mode
-}
-
-// file mode bits
-func (s3f *s3File) Type() fs.FileMode {
-	return s3f.mode
-}
-
-// modification time
-func (s3f *s3File) ModTime() time.Time {
-	return s3f.modTime
-}
-
-// abbreviation for Mode().IsDir()
-func (s3f *s3File) IsDir() bool {
-	return s3f.isDir
-}
-
-// underlying data source (can return nil)
-func (s3f *s3File) Sys() interface{} {
-	return nil
 }
